Standardize parameter key definition of Tuner
TFX users define each module in pipeline through different Key-value pairs.
The raw key string are used in different places. To avoid typo that may break the pipeline, define the key strings as public variables under types, and reuse the variables in other files.

This diff is a cleanup on tuner module.

PiperOrigin-RevId: 354142368